lexbase: introduce EncAlwaysQuoted fmt flag
To accomodate an oracle query, we need to quote the object name regardless if it is
a reserved keyword or is bare identifier. This is because oracle is case insensitive
without quotes. This flag is meant to be used ONLY to construct SELECT stmt for oracle
source.

For example, in oracle:

```
CREATE TABLE "t" ("iD" INT PRIMARY KEY, "AcC" INT)
```

We have to do

```
SELECT * FROM "t"
```

Otherwise we will encounter a "table doesn't exist" error.

Release note: None