lexbase: introduce `EncSkipEscapeString`
EncSkipEscapeString indicates that the string should not be escaped.
More specifically, it means that the tree.DString won't be wrapped with e'text' as the prefix and suffix, and single quotes within the string will not be escaped with a backslash.
This is introduced to for the molt repo, so that the selective data move expression for oracle source won't be wrapped with extra quotes.